feat: add unsafe components
